+++ src/main/java/blackjack/model/Game.java
@@ -2,8 +2,10 @@ package blackjack.model;
 
 import blackjack.model.card.Card;
 import blackjack.model.participant.Dealer;
+import blackjack.model.participant.HitDecisionStrategy;
 import blackjack.model.participant.Participant;
 import blackjack.model.participant.Player;
+import blackjack.model.participant.PlayerHandVisualizer;
 import blackjack.model.participant.Players;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -39,7 +41,8 @@ public class Game {
         }
     }
 
-    private void askHit(Player player, HitDecisionStrategy hitDecisionStrategy, PlayerHandVisualizer playerHandVisualizer) {
+    private void askHit(Player player, HitDecisionStrategy hitDecisionStrategy,
+                        PlayerHandVisualizer playerHandVisualizer) {
         if (player.isBlackjack()) {
             return;
         }
