+++ src/main/java/blackjack/model/participant/HitDecisionStrategy.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.participant;
 
 @FunctionalInterface
 public interface HitDecisionStrategy {
