+++ src/test/java/blackjack/model/GameTest.java
@@ -1,6 +1,5 @@
 package blackjack.model;
 
-import static blackjack.TestFixtures.NO_HIT_STRATEGY;
 import static blackjack.TestFixtures.TEST_EMPTY_VISUALIZER;
 import static blackjack.TestFixtures.createHitDecisionStrategy;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -12,6 +11,7 @@ import blackjack.model.card.Deck;
 import blackjack.model.card.FixedCardShuffler;
 import blackjack.model.card.Suit;
 import blackjack.model.participant.Dealer;
+import blackjack.model.participant.HitDecisionStrategy;
 import blackjack.model.participant.Player;
 import java.util.List;
 import java.util.Map;
